start FxA fixes for new templates
